Attempt to integrate and add configuration.
