+++ pkg/config/config.go
@@ -473,30 +473,12 @@ func initConfig(config Config) {
 	config.SetKnown("apm_config.extra_sample_rate")
 	config.SetKnown("apm_config.dd_agent_bin")
 	config.SetKnown("apm_config.max_events_per_second")
-	config.SetKnown("apm_config.trace_writer.flush_period_seconds")
-	config.SetKnown("apm_config.trace_writer.update_info_period_seconds")
-	config.SetKnown("apm_config.trace_writer.queue.max_age_seconds")
-	config.SetKnown("apm_config.trace_writer.queue.max_bytes")
-	config.SetKnown("apm_config.trace_writer.queue.max_payloads")
-	config.SetKnown("apm_config.trace_writer.queue.exp_backoff_max_duration_seconds")
-	config.SetKnown("apm_config.trace_writer.queue.exp_backoff_base_milliseconds")
-	config.SetKnown("apm_config.trace_writer.queue.exp_backoff_growth_base")
-	config.SetKnown("apm_config.service_writer.flush_period_seconds")
-	config.SetKnown("apm_config.service_writer.update_info_period_seconds")
-	config.SetKnown("apm_config.service_writer.queue.max_age_seconds")
-	config.SetKnown("apm_config.service_writer.queue.max_bytes")
-	config.SetKnown("apm_config.service_writer.queue.max_payloads")
-	config.SetKnown("apm_config.service_writer.queue.exp_backoff_max_duration_seconds")
-	config.SetKnown("apm_config.service_writer.queue.exp_backoff_base_milliseconds")
-	config.SetKnown("apm_config.service_writer.queue.exp_backoff_growth_base")
-	config.SetKnown("apm_config.stats_writer.max_entries_per_payload")
-	config.SetKnown("apm_config.stats_writer.update_info_period_seconds")
-	config.SetKnown("apm_config.stats_writer.queue.max_age_seconds")
-	config.SetKnown("apm_config.stats_writer.queue.max_bytes")
-	config.SetKnown("apm_config.stats_writer.queue.max_payloads")
-	config.SetKnown("apm_config.stats_writer.queue.exp_backoff_max_duration_seconds")
-	config.SetKnown("apm_config.stats_writer.queue.exp_backoff_base_milliseconds")
-	config.SetKnown("apm_config.stats_writer.queue.exp_backoff_growth_base")
+	config.SetKnown("apm_config.trace_writer.connection_limit")
+	config.SetKnown("apm_config.trace_writer.queue_size")
+	config.SetKnown("apm_config.service_writer.connection_limit")
+	config.SetKnown("apm_config.service_writer.queue_size")
+	config.SetKnown("apm_config.stats_writer.connection_limit")
+	config.SetKnown("apm_config.stats_writer.queue_size")
 	config.SetKnown("apm_config.analyzed_rate_by_service.*")
 	config.SetKnown("apm_config.analyzed_spans.*")
