+++ pkg/config/legacy/converter.go
@@ -157,49 +157,30 @@ func FromAgentConfig(agentConfig Config) error {
 
 func extractTraceAgentConfig(agentConfig Config, configConverter *config.LegacyConfigConverter) error {
 	for iniKey, yamlKey := range map[string]string{
-		"trace.api.api_key":                                      "apm_config.api_key",
-		"trace.api.endpoint":                                     "apm_config.apm_dd_url",
-		"trace.config.env":                                       "apm_config.env",
-		"trace.config.log_level":                                 "apm_config.log_level",
-		"trace.config.log_file":                                  "apm_config.log_file",
-		"trace.config.log_throttling":                            "apm_config.log_throttling",
-		"trace.concentrator.bucket_size_seconds":                 "apm_config.bucket_size_seconds",
-		"trace.concentrator.extra_aggregators":                   "apm_config.extra_aggregators",
-		"trace.receiver.receiver_port":                           "apm_config.receiver_port",
-		"trace.receiver.connection_limit":                        "apm_config.connection_limit",
-		"trace.receiver.timeout":                                 "apm_config.receiver_timeout",
-		"trace.sampler.extra_sample_rate":                        "apm_config.extra_sample_rate",
-		"trace.sampler.max_traces_per_second":                    "apm_config.max_traces_per_second",
-		"trace.sampler.max_events_per_second":                    "apm_config.max_events_per_second",
-		"trace.watchdog.max_memory":                              "apm_config.max_memory",
-		"trace.watchdog.max_cpu_percent":                         "apm_config.max_cpu_percent",
-		"trace.watchdog.max_connections":                         "apm_config.max_connections",
-		"trace.watchdog.check_delay_seconds":                     "apm_config.watchdog_check_delay",
-		"trace.writer.services.flush_period_seconds":             "apm_config.service_writer.flush_period_seconds",
-		"trace.writer.services.update_info_period_seconds":       "apm_config.service_writer.update_info_period_seconds",
-		"trace.writer.services.queue_max_age_seconds":            "apm_config.service_writer.queue.max_age_seconds",
-		"trace.writer.services.queue_max_bytes":                  "apm_config.service_writer.queue.max_bytes",
-		"trace.writer.services.queue_max_payloads":               "apm_config.service_writer.queue.max_payloads",
-		"trace.writer.services.exp_backoff_max_duration_seconds": "apm_config.service_writer.queue.exp_backoff_max_duration_seconds",
-		"trace.writer.services.exp_backoff_base_milliseconds":    "apm_config.service_writer.queue.exp_backoff_base_milliseconds",
-		"trace.writer.services.exp_backoff_growth_base":          "apm_config.service_writer.queue.exp_backoff_growth_base",
-		"trace.writer.stats.max_entries_per_payload":             "apm_config.stats_writer.max_entries_per_payload",
-		"trace.writer.stats.update_info_period_seconds":          "apm_config.stats_writer.update_info_period_seconds",
-		"trace.writer.stats.queue_max_age_seconds":               "apm_config.stats_writer.queue.max_age_seconds",
-		"trace.writer.stats.queue_max_bytes":                     "apm_config.stats_writer.queue.max_bytes",
-		"trace.writer.stats.queue_max_payloads":                  "apm_config.stats_writer.queue.max_payloads",
-		"trace.writer.stats.exp_backoff_max_duration_seconds":    "apm_config.stats_writer.queue.exp_backoff_max_duration_seconds",
-		"trace.writer.stats.exp_backoff_base_milliseconds":       "apm_config.stats_writer.queue.exp_backoff_base_milliseconds",
-		"trace.writer.stats.exp_backoff_growth_base":             "apm_config.stats_writer.queue.exp_backoff_growth_base",
-		"trace.writer.traces.max_spans_per_payload":              "apm_config.trace_writer.max_spans_per_payload",
-		"trace.writer.traces.flush_period_seconds":               "apm_config.trace_writer.flush_period_seconds",
-		"trace.writer.traces.update_info_period_seconds":         "apm_config.trace_writer.update_info_period_seconds",
-		"trace.writer.traces.queue_max_age_seconds":              "apm_config.trace_writer.queue.max_age_seconds",
-		"trace.writer.traces.queue_max_bytes":                    "apm_config.trace_writer.queue.max_bytes",
-		"trace.writer.traces.queue_max_payloads":                 "apm_config.trace_writer.queue.max_payloads",
-		"trace.writer.traces.exp_backoff_max_duration_seconds":   "apm_config.trace_writer.queue.exp_backoff_max_duration_seconds",
-		"trace.writer.traces.exp_backoff_base_milliseconds":      "apm_config.trace_writer.queue.exp_backoff_base_milliseconds",
-		"trace.writer.traces.exp_backoff_growth_base":            "apm_config.trace_writer.queue.exp_backoff_growth_base",
+		"trace.api.api_key":                      "apm_config.api_key",
+		"trace.api.endpoint":                     "apm_config.apm_dd_url",
+		"trace.config.env":                       "apm_config.env",
+		"trace.config.log_level":                 "apm_config.log_level",
+		"trace.config.log_file":                  "apm_config.log_file",
+		"trace.config.log_throttling":            "apm_config.log_throttling",
+		"trace.concentrator.bucket_size_seconds": "apm_config.bucket_size_seconds",
+		"trace.concentrator.extra_aggregators":   "apm_config.extra_aggregators",
+		"trace.receiver.receiver_port":           "apm_config.receiver_port",
+		"trace.receiver.connection_limit":        "apm_config.connection_limit",
+		"trace.receiver.timeout":                 "apm_config.receiver_timeout",
+		"trace.sampler.extra_sample_rate":        "apm_config.extra_sample_rate",
+		"trace.sampler.max_traces_per_second":    "apm_config.max_traces_per_second",
+		"trace.sampler.max_events_per_second":    "apm_config.max_events_per_second",
+		"trace.watchdog.max_memory":              "apm_config.max_memory",
+		"trace.watchdog.max_cpu_percent":         "apm_config.max_cpu_percent",
+		"trace.watchdog.max_connections":         "apm_config.max_connections",
+		"trace.watchdog.check_delay_seconds":     "apm_config.watchdog_check_delay",
+		"trace.writer.services.connection_limit": "apm_config.service_writer.connection_limit",
+		"trace.writer.services.queue_size":       "apm_config.service_writer.queue_size",
+		"trace.writer.stats.connection_limit":    "apm_config.stats_writer.connection_limit",
+		"trace.writer.stats.queue_size":          "apm_config.stats_writer.queue_size",
+		"trace.writer.traces.connection_limit":   "apm_config.trace_writer.connection_limit",
+		"trace.writer.traces.queue_size":         "apm_config.trace_writer.queue_size",
 	} {
 		if v, ok := agentConfig[iniKey]; ok {
 			configConverter.Set(yamlKey, v)
