+++ pkg/config/legacy/tests/datadog.conf
@@ -317,35 +317,16 @@ max_connections: 40
 check_delay_seconds: 5
 
 [trace.writer.services]
-flush_period_seconds: 1
-update_info_period_seconds: 1
-queue_max_age_seconds: 1
-queue_max_bytes: 456
-queue_max_payloads: 4
-exp_backoff_max_duration_seconds: 4
-exp_backoff_base_milliseconds: 1
-exp_backoff_growth_base: 2
+connection_limit: 1
+queue_size: 2
 
 [trace.writer.stats]
-max_entries_per_payload: 10
-update_info_period_seconds: 2
-queue_max_age_seconds: 1
-queue_max_bytes: 456
-queue_max_payloads: 4
-exp_backoff_max_duration_seconds: 4
-exp_backoff_base_milliseconds: 1
-exp_backoff_growth_base: 2
+connection_limit: 3
+queue_size: 4
 
 [trace.writer.traces]
-max_spans_per_payload: 100
-flush_period_seconds: 3
-update_info_period_seconds: 2
-queue_max_age_seconds: 1
-queue_max_bytes: 456 # 64-bit integer
-queue_max_payloads: 4
-exp_backoff_max_duration_seconds: 4
-exp_backoff_base_milliseconds: 1
-exp_backoff_growth_base: 2
+connection_limit: 5
+queue_size: 6
 
 [trace.ignore]
 # a blacklist of regular expressions can be provided to disable certain traces based on their resource name
