+++ pkg/trace/writer/sender.go
@@ -1,383 +1,361 @@
 package writer
 
 import (
-	"container/list"
+	"bytes"
+	"errors"
 	"fmt"
+	"math"
+	"math/rand"
+	"net/http"
+	"net/url"
+	"strconv"
 	"sync"
+	"sync/atomic"
 	"time"
 
-	"github.com/DataDog/datadog-agent/pkg/trace/watchdog"
-	"github.com/DataDog/datadog-agent/pkg/trace/writer/backoff"
-	writerconfig "github.com/DataDog/datadog-agent/pkg/trace/writer/config"
+	"github.com/DataDog/datadog-agent/pkg/trace/config"
+	"github.com/DataDog/datadog-agent/pkg/trace/info"
+	"github.com/DataDog/datadog-agent/pkg/trace/osutil"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
-// payload represents a data payload to be sent to some endpoint
-type payload struct {
-	creationDate time.Time
-	bytes        []byte
-	headers      map[string]string
-}
-
-// newPayload constructs a new payload object with the provided data and with CreationDate initialized to the current
-// time.
-func newPayload(bytes []byte, headers map[string]string) *payload {
-	return &payload{
-		creationDate: time.Now(),
-		bytes:        bytes,
-		headers:      headers,
+// newSenders returns a list of senders based on the given agent configuration, using climit
+// as the maximum number of concurrent outgoing connections, writing to path. The given
+// namespace is used as a prefix for reported metrics.
+func newSenders(cfg *config.AgentConfig, r eventRecorder, path string, climit, qsize int) []*sender {
+	if e := cfg.Endpoints; len(e) == 0 || e[0].Host == "" || e[0].APIKey == "" {
+		panic(errors.New("config was not properly validated"))
 	}
+	client := cfg.HTTPClient()
+	// spread out the the maximum connection limit (climit) between senders
+	maxConns := math.Max(1, float64(climit/len(cfg.Endpoints)))
+	senders := make([]*sender, len(cfg.Endpoints))
+	for i, endpoint := range cfg.Endpoints {
+		url, err := url.Parse(endpoint.Host + path)
+		if err != nil {
+			osutil.Exitf("Invalid host endpoint: %q", endpoint.Host)
+		}
+		senders[i] = newSender(&senderConfig{
+			client:    client,
+			maxConns:  int(maxConns),
+			maxQueued: qsize,
+			url:       url,
+			apiKey:    endpoint.APIKey,
+			recorder:  r,
+		})
+	}
+	return senders
 }
 
-// eventType is a type of event sent down the monitor channel.
+// eventRecorder implementations are able to take note of events happening in
+// the sender.
+type eventRecorder interface {
+	// recordEvent notifies that event t has happened, passing details about
+	// the event in data.
+	recordEvent(t eventType, data *eventData)
+}
+
+// eventType specifies an event which occurred in the sender.
 type eventType int
 
 const (
-	eventTypeSuccess eventType = iota
-	eventTypeFailure
-	eventTypeRetry
+	// eventTypeRetry specifies that a send failed with a retriable error (5xx).
+	eventTypeRetry eventType = iota
+	// eventTypeSent specifies that a single payload was successfully sent.
+	eventTypeSent
+	// eventTypeRejected specifies that the edge rejected this payload.
+	eventTypeRejected
+	// eventTypeDropped specifies that a payload had to be dropped to make room
+	// in the queue.
+	eventTypeDropped
 )
 
 var eventTypeStrings = map[eventType]string{
-	eventTypeSuccess: "success",
-	eventTypeFailure: "failure",
-	eventTypeRetry:   "retry",
+	eventTypeRetry:    "eventTypeRetry",
+	eventTypeSent:     "eventTypeSent",
+	eventTypeRejected: "eventTypeRejected",
+	eventTypeDropped:  "eventTypeDropped",
 }
 
-func (e eventType) String() string { return eventTypeStrings[e] }
+// String implements fmt.Stringer.
+func (t eventType) String() string { return eventTypeStrings[t] }
 
-type monitorEvent struct {
-	typ        eventType
-	payload    *payload
-	stats      sendStats
-	err        error
-	retryDelay time.Duration
-	retryNum   int
+// eventData represents information about a sender event. Not all fields apply
+// to all events.
+type eventData struct {
+	// host specifies the host which the sender is sending to.
+	host string
+	// bytes represents the number of bytes affected by this event.
+	bytes int
+	// count specfies the number of payloads that this events refers to.
+	count int
+	// duration specifies the time it took to complete this event. It
+	// is set for eventType{Sent,Retry,Rejected}.
+	duration time.Duration
+	// err specifies the error that may have occurred on events eventType{Retry,Rejected}.
+	err error
+	// connectionFill specifies the percentage of allowed connections used.
+	// At 100% (1.0) the writer will become blocking.
+	connectionFill float64
+	// queueFill specifies how flul the queue is. It's a floating point number ranging
+	// between 0 (0%) and 1 (100%).
+	queueFill float64
 }
 
-// sendStats represents basic stats related to the sending of a payload.
-type sendStats struct {
-	sendTime time.Duration
-	host     string
+// senderConfig specifies the configuration for the sender.
+type senderConfig struct {
+	// client specifies the HTTP client to use when sending requests.
+	client *http.Client
+	// url specifies the URL to send requests too.
+	url *url.URL
+	// apiKey specifies the Datadog API key to use.
+	apiKey string
+	// maxConns specifies the maximum number of allowed concurrent ougoing
+	// connections.
+	maxConns int
+	// maxQueued specifies the maximum number of payloads allowed in the queue.
+	// When it is surpassed, oldest items get dropped to make room for new ones.
+	maxQueued int
+	// recorder specifies the eventRecorder to use when reporting events occurring
+	// in the sender.
+	recorder eventRecorder
 }
 
-// payloadSender represents an object capable of asynchronously sending payloads to some endpoint.
-type payloadSender interface {
-	Start()
-	Run()
-	Stop()
-	Send(payload *payload)
-	Monitor() <-chan monitorEvent
+// sender is responsible for sending payloads to a given URL. It uses a size-limited
+// retry queue with a backoff mechanism in case of retriable errors.
+type sender struct {
+	cfg *senderConfig
 
-	setEndpoint(endpoint)
+	queue    chan *payload // payload queue
+	climit   chan struct{} // semaphore for limiting concurrent connections
+	inflight int32         // inflight payloads
+	attempt  int32         // active retry attempt
 }
 
-// queuableSender is a specific implementation of a payloadSender that will queue new payloads on error and
-// retry sending them according to some configurable BackoffTimer.
-type queuableSender struct {
-	conf writerconfig.QueuablePayloadSenderConf
-
-	mu                sync.RWMutex // guards below group
-	queuedPayloads    *list.List
-	currentQueuedSize int64
-	backoffTimer      backoff.Timer
-
-	syncBarrier <-chan interface{} // used only in tests
-	wg          sync.WaitGroup     // tracks active sends
-
-	in        chan *payload
-	monitorCh chan monitorEvent
-	endpoint  endpoint
-
-	exit chan struct{}
+// newSender returns a new sender based on the given config cfg.
+func newSender(cfg *senderConfig) *sender {
+	s := sender{
+		cfg:    cfg,
+		queue:  make(chan *payload, cfg.maxQueued),
+		climit: make(chan struct{}, cfg.maxConns),
+	}
+	go s.loop()
+	return &s
 }
 
-// newSender constructs a new QueuablePayloadSender with custom configuration to send payloads to
-// the provided endpoint.
-func newSender(e endpoint, conf writerconfig.QueuablePayloadSenderConf) *queuableSender {
-	return &queuableSender{
-		conf:           conf,
-		queuedPayloads: list.New(),
-		backoffTimer:   backoff.NewCustomExponentialTimer(conf.ExponentialBackoff),
-		in:             make(chan *payload, conf.InChannelSize),
-		monitorCh:      make(chan monitorEvent),
-		endpoint:       e,
-		exit:           make(chan struct{}),
+// loop runs the main sender loop.
+func (s *sender) loop() {
+	for p := range s.queue {
+		s.backoff()
+		s.climit <- struct{}{}
+		go func(p *payload) {
+			defer func() { <-s.climit }()
+			s.sendPayload(p)
+		}(p)
 	}
 }
 
-// Send sends a single isolated payload through this sender.
-func (s *queuableSender) Send(payload *payload) {
-	s.in <- payload
-}
-
-// Stop asks this sender to stop and waits until it correctly stops.
-func (s *queuableSender) Stop() {
-	close(s.in)
-	<-s.exit
-	s.wg.Wait()
-	close(s.monitorCh)
-}
-
-func (s *queuableSender) setEndpoint(e endpoint) {
-	s.endpoint = e
-}
-
-// Monitor allows an external entity to monitor events of this sender by receiving Sender*Event structs.
-func (s *queuableSender) Monitor() <-chan monitorEvent {
-	return s.monitorCh
-}
-
-// send will send the provided payload without any checks.
-func (s *queuableSender) doSend(payload *payload) (sendStats, error) {
-	if payload == nil {
-		return sendStats{}, nil
-	}
-
-	startFlush := time.Now()
-	err := s.endpoint.write(payload)
-
-	sendStats := sendStats{
-		sendTime: time.Since(startFlush),
-		host:     s.endpoint.baseURL(),
-	}
-
-	return sendStats, err
-}
-
-// Start asynchronously starts this QueueablePayloadSender.
-func (s *queuableSender) Start() {
-	go func() {
-		defer watchdog.LogOnPanic()
-		s.Run()
-	}()
-}
-
-// Run executes the queuableSender main logic synchronously.
-func (s *queuableSender) Run() {
-	defer close(s.exit)
-	sema := make(chan struct{}, s.conf.MaxConnections)
-
-	for {
-		select {
-		case payload, more := <-s.in:
-			if !more {
-				log.Info("Exiting payload sender, try flushing whatever is left")
-				s.flushQueue()
-				return
-			}
-			if payload == nil {
-				continue
-			}
-			s.mu.RLock()
-			queuing := s.queuedPayloads.Len() > 0
-			s.mu.RUnlock()
-			if queuing {
-				if err := s.enqueue(payload); err != nil {
-					log.Debugf("Error while queueing payload: %v", err)
-					s.notifyError(payload, err)
-				}
-				continue
-			}
-			sema <- struct{}{}
-			s.wg.Add(1)
-			go func() {
-				defer s.wg.Done()
-				if err := s.trySend(payload); err != nil {
-					log.Debugf("Error while sending or queueing payload. err=%v", err)
-					s.notifyError(payload, err)
-				}
-				<-sema
-			}()
-		case <-s.backoffTimer.ReceiveTick():
-			s.flushQueue()
-		case <-s.syncBarrier:
-			// TODO: Is there a way of avoiding this? I want Promises in Go :(((
-			// This serves as a barrier (assuming syncBarrier is an unbuffered channel). Used for testing
-			continue
-		}
-	}
-}
-
-// trySend sends the provided payload or queues it if this sender is currently queueing payloads.
-func (s *queuableSender) trySend(payload *payload) error {
-	stats, err := s.doSend(payload)
-	if err != nil {
-		if _, ok := err.(*retriableError); ok {
-			// start a retry queue
-			s.mu.Lock()
-			defer s.mu.Unlock()
-			retryNum, delay := s.backoffTimer.ScheduleRetry(err)
-			log.Debugf("Got retriable error. Starting a queue. delay=%s, err=%v", delay, err)
-			s.notifyRetry(payload, err, delay, retryNum)
-			return s.enqueueLocked(payload)
-		}
-		return err
-	}
-	log.Tracef("Successfully sent direct payload: %v", payload)
-	s.notifySuccess(payload, stats)
-	return nil
-}
-
-func (s *queuableSender) enqueue(p *payload) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.enqueueLocked(p)
-}
-
-func (s *queuableSender) enqueueLocked(payload *payload) error {
-	// Start by discarding payloads that are too old, freeing up memory
-	s.discardOldPayloads()
-
-	for s.conf.MaxQueuedPayloads > 0 && s.queuedPayloads.Len() >= s.conf.MaxQueuedPayloads {
-		log.Debugf("Dropping existing payload because max queued payloads reached: %d", s.conf.MaxQueuedPayloads)
-		if _, err := s.dropOldestPayload("max queued payloads reached"); err != nil {
-			panic(fmt.Errorf("unable to respect max queued payloads value of %d", s.conf.MaxQueuedPayloads))
-		}
-	}
-
-	newPayloadSize := int64(len(payload.bytes))
-
-	if s.conf.MaxQueuedBytes > 0 && newPayloadSize > s.conf.MaxQueuedBytes {
-		log.Debugf("Payload bigger than max size: size=%d, max size=%d", newPayloadSize, s.conf.MaxQueuedBytes)
-		return fmt.Errorf("unable to queue payload bigger than max size: payload size=%d, max size=%d",
-			newPayloadSize, s.conf.MaxQueuedBytes)
-	}
-
-	for s.conf.MaxQueuedBytes > 0 && s.currentQueuedSize+newPayloadSize > s.conf.MaxQueuedBytes {
-		if _, err := s.dropOldestPayload("max queued bytes reached"); err != nil {
-			// Should never happen because we know we can fit it in
-			panic(fmt.Errorf("unable to find space for queueing payload of size %d: %v", newPayloadSize, err))
-		}
-	}
-
-	log.Tracef("Queuing new payload: %v", payload)
-	s.queuedPayloads.PushBack(payload)
-	s.currentQueuedSize += newPayloadSize
-
-	return nil
-}
-
-func (s *queuableSender) flushQueue() error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	log.Debugf("Attempting to flush queue with %d payloads", s.queuedPayloads.Len())
-
-	// Start by discarding payloads that are too old
-	s.discardOldPayloads()
-
-	// For the remaining ones, try to send them one by one
-	var next *list.Element
-	for e := s.queuedPayloads.Front(); e != nil; e = next {
-		payload := e.Value.(*payload)
-
-		var err error
-		var stats sendStats
-
-		if stats, err = s.doSend(payload); err != nil {
-			if _, ok := err.(*retriableError); ok {
-				// If send failed due to a retriable error, retry flush later
-				retryNum, delay := s.backoffTimer.ScheduleRetry(err)
-				log.Debugf("Got retriable error. Retrying flush later: retry=%d, delay=%s, err=%v",
-					retryNum, delay, err)
-				s.notifyRetry(payload, err, delay, retryNum)
-				// Don't try to send following. We'll flush all later.
-				return err
-			}
-
-			// If send failed due to non-retriable error, notify error and drop it
-			log.Debugf("Dropping payload due to non-retriable error: err=%v, payload=%v", err, payload)
-			s.notifyError(payload, err)
-			next = s.removeQueuedPayload(e)
-			// Try sending next ones
-			continue
-		}
-
-		// If successful, remove payload from queue
-		log.Tracef("Successfully sent a queued payload: %v", payload)
-		s.notifySuccess(payload, stats)
-		next = s.removeQueuedPayload(e)
-	}
-
-	s.backoffTimer.Reset()
-
-	return nil
-}
-
-func (s *queuableSender) removeQueuedPayload(e *list.Element) *list.Element {
-	next := e.Next()
-	payload := e.Value.(*payload)
-	s.currentQueuedSize -= int64(len(payload.bytes))
-	s.queuedPayloads.Remove(e)
-	return next
-}
-
-// Discard those payloads that are older than max age.
-func (s *queuableSender) discardOldPayloads() {
-	// If MaxAge <= 0 then age limitation is disabled so do nothing
-	if s.conf.MaxAge <= 0 {
+// backoff triggers a sleep period proportional to the retry attempt, if any.
+func (s *sender) backoff() {
+	attempt := atomic.LoadInt32(&s.attempt)
+	delay := backoffDuration(int(attempt))
+	if delay == 0 {
 		return
 	}
+	time.Sleep(delay)
+}
 
-	var next *list.Element
-
-	for e := s.queuedPayloads.Front(); e != nil; e = next {
-		payload := e.Value.(*payload)
-
-		age := time.Since(payload.creationDate)
-
-		// Payloads are kept in order so as soon as we find one that isn't, we can break out
-		if age < s.conf.MaxAge {
-			break
+// Stop stops the sender. It attempts to wait for all inflight payloads to complete
+// with a timeout of 5 seconds.
+func (s *sender) Stop() {
+	timeout := time.After(5 * time.Second)
+outer:
+	for {
+		select {
+		case <-timeout:
+			break outer
+		default:
+			if atomic.LoadInt32(&s.inflight) == 0 {
+				break outer
+			}
+			time.Sleep(10 * time.Millisecond)
 		}
+	}
+	close(s.queue)
+}
 
-		err := fmt.Errorf("payload is older than max age: age=%v, max age=%v", age, s.conf.MaxAge)
-		log.Tracef("Discarding payload: err=%v, payload=%v", err, payload)
-		s.notifyError(payload, err)
-		next = s.removeQueuedPayload(e)
+// Push pushes p onto the sender's queue, to be written to the destination.
+func (s *sender) Push(p *payload) {
+	for {
+		select {
+		case s.queue <- p:
+			// ok
+			atomic.AddInt32(&s.inflight, 1)
+			return
+		default:
+			// drop the oldest item in the queue to make room
+			select {
+			case drop := <-s.queue:
+				s.recordEvent(eventTypeDropped, &eventData{
+					bytes: drop.body.Len(),
+					count: 1,
+				})
+			default:
+				// the queue got drained; not very likely to happen, but
+				// we shouldn't risk a deadlock
+				continue
+			}
+		}
 	}
 }
 
-// Payloads are kept in order so dropping the one at the front guarantees we're dropping the oldest
-func (s *queuableSender) dropOldestPayload(reason string) (*payload, error) {
-	if s.queuedPayloads.Len() == 0 {
-		return nil, fmt.Errorf("no queued payloads")
+// sendPayload sends the payload p to the destination URL.
+func (s *sender) sendPayload(p *payload) {
+	req, err := p.httpRequest(s.cfg.url)
+	if err != nil {
+		log.Errorf("http.Request: %s", err)
+		return
 	}
-
-	err := fmt.Errorf("payload dropped: %s", reason)
-	droppedPayload := s.queuedPayloads.Front().Value.(*payload)
-	s.removeQueuedPayload(s.queuedPayloads.Front())
-	s.notifyError(droppedPayload, err)
-
-	return droppedPayload, nil
+	start := time.Now()
+	err = s.do(req)
+	stats := &eventData{
+		bytes:    p.body.Len(),
+		count:    1,
+		duration: time.Since(start),
+		err:      err,
+	}
+	switch err.(type) {
+	case *retriableError:
+		// request failed again, but can be retried
+		atomic.AddInt32(&s.attempt, 1)
+		select {
+		case s.queue <- p:
+			s.recordEvent(eventTypeRetry, stats)
+			return
+		default:
+			// queue is full; since this is the oldest payload, we drop it
+			s.recordEvent(eventTypeDropped, stats)
+		}
+	case nil:
+		// request was successful
+		atomic.SwapInt32(&s.attempt, 0)
+		s.recordEvent(eventTypeSent, stats)
+	default:
+		// this is a fatal error, we have to drop this payload
+		s.recordEvent(eventTypeRejected, stats)
+	}
+	// we're done with p
+	atomic.AddInt32(&s.inflight, -1)
 }
 
-func (s *queuableSender) notifySuccess(payload *payload, sendStats sendStats) {
-	s.sendEvent(&monitorEvent{
-		typ:     eventTypeSuccess,
-		payload: payload,
-		stats:   sendStats,
-	})
+// recordEvent records the occurrence of the given event type t. It additionally
+// passes on the data and augments it with additional information.
+func (s *sender) recordEvent(t eventType, data *eventData) {
+	if recorder := s.cfg.recorder; recorder != nil {
+		data.host = s.cfg.url.Hostname()
+		data.connectionFill = float64(len(s.climit)) / float64(cap(s.climit))
+		data.queueFill = float64(len(s.queue)) / float64(cap(s.queue))
+		recorder.recordEvent(t, data)
+	}
 }
 
-func (s *queuableSender) notifyError(payload *payload, err error) {
-	s.sendEvent(&monitorEvent{
-		typ:     eventTypeFailure,
-		payload: payload,
-		err:     err,
-	})
+// userAgent is the computed user agent we'll use when communicating with Datadog var
+var userAgent = fmt.Sprintf("Datadog Trace Agent/%s/%s", info.Version, info.GitCommit)
+
+// retriableError is an error returned by the server which may be retried at a later time.
+type retriableError struct{ err error }
+
+// Error implements error.
+func (e retriableError) Error() string { return e.err.Error() }
+
+const (
+	headerAPIKey    = "DD-Api-Key"
+	headerUserAgent = "User-Agent"
+)
+
+func (s *sender) do(req *http.Request) error {
+	req.Header.Set(headerAPIKey, s.cfg.apiKey)
+	req.Header.Set(headerUserAgent, userAgent)
+	resp, err := s.cfg.client.Do(req)
+	if err != nil {
+		// request errors are either redirect errors, url errors
+		return &retriableError{err}
+	}
+	if resp.StatusCode/100 == 5 {
+		// 5xx errors can be retried
+		return &retriableError{
+			fmt.Errorf("server responded with %q", resp.Status),
+		}
+	}
+	if resp.StatusCode/100 != 2 {
+		// non-2xx errors are failures
+		return errors.New(resp.Status)
+	}
+	return nil
 }
 
-func (s *queuableSender) notifyRetry(payload *payload, err error, delay time.Duration, retryNum int) {
-	s.sendEvent(&monitorEvent{
-		typ:        eventTypeRetry,
-		payload:    payload,
-		err:        err,
-		retryDelay: delay,
-		retryNum:   retryNum,
-	})
+// payloads specifies a payload to be sent by the sender.
+type payload struct {
+	body    *bytes.Buffer     // request body
+	headers map[string]string // request headers
 }
 
-func (s *queuableSender) sendEvent(event *monitorEvent) {
-	s.monitorCh <- *event
+// newPayload creates a new payload with the given headers.
+func newPayload(headers map[string]string) *payload {
+	p := &payload{
+		body:    new(bytes.Buffer),
+		headers: make(map[string]string),
+	}
+	for k, v := range headers {
+		p.headers[k] = v
+	}
+	return p
+}
+
+// httpRequest returns an HTTP request based on the payload, targeting the given URL.
+func (p *payload) httpRequest(url *url.URL) (*http.Request, error) {
+	req, err := http.NewRequest(http.MethodPost, url.String(), bytes.NewReader(p.body.Bytes()))
+	if err != nil {
+		// this should never happen with sanitized data (invalid method or invalid url)
+		return nil, err
+	}
+	for k, v := range p.headers {
+		req.Header.Add(k, v)
+	}
+	req.Header.Add("Content-Length", strconv.Itoa(p.body.Len()))
+	return req, nil
+}
+
+// stopSenders attempts to simultaneously stop a group of senders.
+func stopSenders(senders []*sender) {
+	var wg sync.WaitGroup
+	for _, s := range senders {
+		wg.Add(1)
+		go func(s *sender) {
+			defer wg.Done()
+			s.Stop()
+		}(s)
+	}
+	wg.Wait()
+}
+
+const (
+	// backoffBase specifies the multiplier base for the backoff duration algorithm.
+	backoffBase = 100 * time.Millisecond
+	// backoffMaxDuration is the maximum permitted backoff duration.
+	backoffMaxDuration = 10 * time.Second
+)
+
+// backoffDuration returns the backoff duration necessary for the given attempt.
+// The formula is "Full Jitter":
+//   random_between(0, min(cap, base * 2 ** attempt))
+// https://aws.amazon.com/blogs/architecture/exponential-backoff-and-jitter/
+var backoffDuration = func(attempt int) time.Duration {
+	if attempt == 0 {
+		return 0
+	}
+	maxPow := float64(backoffMaxDuration / backoffBase)
+	pow := math.Min(math.Pow(2, float64(attempt)), maxPow)
+	ns := int64(float64(backoffBase) * pow)
+	return time.Duration(rand.Int63n(ns))
 }
