+++ pkg/trace/writer/trace.go
@@ -1,8 +1,8 @@
 package writer
 
 import (
-	"bytes"
 	"compress/gzip"
+	"math"
 	"strings"
 	"sync/atomic"
 	"time"
@@ -13,290 +13,226 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/trace/metrics/timing"
 	"github.com/DataDog/datadog-agent/pkg/trace/pb"
 	"github.com/DataDog/datadog-agent/pkg/trace/traceutil"
-	"github.com/DataDog/datadog-agent/pkg/trace/watchdog"
-	writerconfig "github.com/DataDog/datadog-agent/pkg/trace/writer/config"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
-	"github.com/golang/protobuf/proto"
+
+	"github.com/gogo/protobuf/proto"
 )
 
+// pathTraces is the target host API path for delivering traces.
 const pathTraces = "/api/v0.2/traces"
 
-// payloadFlushThreshold specifies the maximum accumulated payload size that is allowed before
+// maxPayloadSize specifies the maximum accumulated payload size that is allowed before
 // a flush is triggered; replaced in tests.
-var payloadFlushThreshold = 3200000 // 3.2MB is the maximum allowed by the Datadog API
+var maxPayloadSize = 3200000 // 3.2MB is the maximum allowed by the Datadog API
 
-// TracePackage represents the result of a trace sampling operation.
-//
-// NOTE: A TracePackage can be valid even if any of its fields is nil/empty. In particular, a common case is that of
-// empty Trace but non-empty Events. This happens when events are extracted from a trace that wasn't sampled.
-type TracePackage struct {
+// SampledSpans represents the result of a trace sampling operation.
+type SampledSpans struct {
 	// Trace will contain a trace if it was sampled or be empty if it wasn't.
 	Trace pb.Trace
-	// Events contains all APMEvents extracted from a trace. If no events were extracted, it will be empty.
+	// Events contains all APM events extracted from a trace. If no events were extracted, it will be empty.
 	Events []*pb.Span
 }
 
 // Empty returns true if this TracePackage has no data.
-func (p *TracePackage) Empty() bool {
-	return len(p.Trace) == 0 && len(p.Events) == 0
+func (ss *SampledSpans) Empty() bool {
+	return len(ss.Trace) == 0 && len(ss.Events) == 0
 }
 
 // size returns the estimated size of the package.
-func (p *TracePackage) size() int {
+func (ss *SampledSpans) size() int {
 	// we use msgpack's Msgsize() heuristic because it is a good indication
 	// of the weight of a span and the msgpack size is relatively close to
 	// the protobuf size, which is expensive to compute.
-	return p.Trace.Msgsize() + pb.Trace(p.Events).Msgsize()
+	return ss.Trace.Msgsize() + pb.Trace(ss.Events).Msgsize()
 }
 
-// TraceWriter ingests sampled traces and flushes them to the API.
+// TraceWriter buffers traces and APM events, flushing them to the Datadog API.
 type TraceWriter struct {
-	stats    info.TraceWriterInfo
-	hostName string
+	in       <-chan *SampledSpans
+	hostname string
 	env      string
-	conf     writerconfig.TraceWriterConfig
-	in       <-chan *TracePackage
+	senders  []*sender
+	stop     chan struct{}
+	stats    *info.TraceWriterInfo
 
-	traces        []*pb.APITrace
-	events        []*pb.Span
-	spansInBuffer int
-	bytesInBuffer int // estimated size of next payload
-
-	sender payloadSender
-	exit   chan struct{}
+	traces       []*pb.APITrace // traces buffered
+	events       []*pb.Span     // events buffered
+	bufferedSize int            // estimated buffer size
 }
 
-// NewTraceWriter returns a new writer for traces.
-func NewTraceWriter(conf *config.AgentConfig, in <-chan *TracePackage) *TraceWriter {
-	cfg := conf.TraceWriterConfig
-	endpoints := newEndpoints(conf, pathTraces)
-	sender := newMultiSender(endpoints, cfg.SenderConfig)
-	log.Infof("Trace writer initializing with config: %+v", cfg)
-
-	return &TraceWriter{
-		conf:     cfg,
-		hostName: conf.Hostname,
-		env:      conf.DefaultEnv,
-
-		traces: []*pb.APITrace{},
-		events: []*pb.Span{},
-
-		in: in,
-
-		sender: sender,
-		exit:   make(chan struct{}),
+// NewTraceWriter returns a new TraceWriter. It is created for the given agent configuration and
+// will accept incoming spans via the in channel.
+func NewTraceWriter(cfg *config.AgentConfig, in <-chan *SampledSpans) *TraceWriter {
+	tw := &TraceWriter{
+		in:       in,
+		hostname: cfg.Hostname,
+		env:      cfg.DefaultEnv,
+		stats:    &info.TraceWriterInfo{},
+		stop:     make(chan struct{}),
 	}
+	climit := cfg.TraceWriter.ConnectionLimit
+	if climit == 0 {
+		// default to 10% of the connection limit to outgoing sends.
+		climit = int(math.Max(1, float64(cfg.ConnectionLimit)/10))
+	}
+	qsize := cfg.TraceWriter.QueueSize
+	if qsize == 0 {
+		// default to 50% of maximum memory.
+		qsize = int(math.Max(1, cfg.MaxMemory/2/float64(maxPayloadSize)))
+	}
+	log.Debugf("Trace writer initialized (climit=%d qsize=%d)", climit, qsize)
+	tw.senders = newSenders(cfg, tw, pathTraces, climit, qsize)
+	return tw
 }
 
-// Start starts the writer.
-func (w *TraceWriter) Start() {
-	w.sender.Start()
-	go func() {
-		defer watchdog.LogOnPanic()
-		w.Run()
-	}()
+// Stop stops the TraceWriter and attempts to flush whatever is left in the senders buffers.
+func (w *TraceWriter) Stop() {
+	log.Debug("Exiting trace writer. Trying to flush whatever is left...")
+	w.stop <- struct{}{}
+	<-w.stop
+	stopSenders(w.senders)
 }
 
-// Run runs the main loop of the writer goroutine. It sends traces to the payload constructor, flushing it periodically
-// and collects stats which are also reported periodically.
+// Run starts the TraceWriter.
 func (w *TraceWriter) Run() {
-	defer close(w.exit)
-
-	// for now, simply flush every x seconds
-	flushTicker := time.NewTicker(w.conf.FlushPeriod)
-	defer flushTicker.Stop()
-
-	updateInfoTicker := time.NewTicker(w.conf.UpdateInfoPeriod)
-	defer updateInfoTicker.Stop()
-
-	// Monitor sender for events
-	go func() {
-		for event := range w.sender.Monitor() {
-			switch event.typ {
-			case eventTypeSuccess:
-				log.Debugf("Flushed trace payload to the API, time:%s, size:%d bytes", event.stats.sendTime,
-					len(event.payload.bytes))
-				tags := []string{"url:" + event.stats.host}
-				metrics.Gauge("datadog.trace_agent.trace_writer.flush_duration",
-					event.stats.sendTime.Seconds(), tags, 1)
-				atomic.AddInt64(&w.stats.Payloads, 1)
-			case eventTypeFailure:
-				log.Errorf("Failed to flush trace payload, host:%s, time:%s, size:%d bytes, error: %s",
-					event.stats.host, event.stats.sendTime, len(event.payload.bytes), event.err)
-				atomic.AddInt64(&w.stats.Errors, 1)
-			case eventTypeRetry:
-				log.Errorf("Retrying flush trace payload, retryNum: %d, delay:%s, error: %s",
-					event.retryNum, event.retryDelay, event.err)
-				atomic.AddInt64(&w.stats.Retries, 1)
-			default:
-				log.Debugf("Unable to handle event with type %T", event)
-			}
-		}
-	}()
-
-	log.Debug("Starting trace writer")
-
+	t := time.NewTicker(5 * time.Second)
+	defer t.Stop()
+	defer close(w.stop)
 	for {
 		select {
-		case sampledTrace := <-w.in:
-			w.handleSampledTrace(sampledTrace)
-		case <-flushTicker.C:
-			log.Debug("Flushing current traces")
+		case pkg := <-w.in:
+			w.addSpans(pkg)
+		case <-w.stop:
+			// drain the input channel before stopping
+		outer:
+			for {
+				select {
+				case pkg := <-w.in:
+					w.addSpans(pkg)
+				default:
+					break outer
+				}
+			}
 			w.flush()
-		case <-updateInfoTicker.C:
-			go w.updateInfo()
-		case <-w.exit:
-			log.Info("Exiting trace writer, flushing all remaining traces")
-			w.flush()
-			w.updateInfo()
-			log.Info("Flushed. Exiting")
 			return
+		case <-t.C:
+			w.report()
+			w.flush()
 		}
 	}
 }
 
-// Stop stops the main Run loop.
-func (w *TraceWriter) Stop() {
-	w.exit <- struct{}{}
-	<-w.exit
-	w.sender.Stop()
-}
-
-func (w *TraceWriter) handleSampledTrace(pkg *TracePackage) {
-	if pkg == nil || pkg.Empty() {
-		log.Debug("Ignoring empty sampled trace")
+func (w *TraceWriter) addSpans(pkg *SampledSpans) {
+	if pkg.Empty() {
 		return
 	}
 
-	size := pkg.size()
+	atomic.AddInt64(&w.stats.Spans, int64(len(pkg.Trace)))
+	atomic.AddInt64(&w.stats.Traces, 1)
+	atomic.AddInt64(&w.stats.Events, int64(len(pkg.Events)))
 
-	if w.bytesInBuffer > 0 && w.bytesInBuffer+size > payloadFlushThreshold {
-		// this new package would push us over, flush
-		log.Debug("Flushing... Reached size threshold.")
+	size := pkg.size()
+	if size+w.bufferedSize > maxPayloadSize {
+		// reached maximum allowed buffered size
 		w.flush()
 	}
-
 	if len(pkg.Trace) > 0 {
-		log.Tracef("Handling new trace with %d spans: %v", len(pkg.Trace), pkg.Trace)
 		w.traces = append(w.traces, traceutil.APITrace(pkg.Trace))
 	}
-	if len(pkg.Events) > 0 {
-		log.Tracef("Handling new APM events: %v", pkg.Events)
-		w.events = append(w.events, pkg.Events...)
-	}
-
-	w.bytesInBuffer += size
-	w.spansInBuffer += len(pkg.Trace) + len(pkg.Events)
-
-	if size > payloadFlushThreshold {
-		// we've added a single package that surpasses our threshold, we should count this occurrence,
-		// it could be an indication of "over instrumentation" on the client side, where too many spans
-		// could be present in traces.
-		atomic.AddInt64(&w.stats.SingleMaxSize, 1)
-		log.Debugf("Flushing... Surpassed size threshold with a single package sized %d bytes", size)
-		w.flush()
-	}
+	w.events = append(w.events, pkg.Events...)
+	w.bufferedSize += size
 }
 
-func (w *TraceWriter) flush() {
-	numTraces := len(w.traces)
-	numEvents := len(w.events)
+func (w *TraceWriter) resetBuffer() {
+	w.bufferedSize = 0
+	w.traces = w.traces[:0]
+	w.events = w.events[:0]
+}
 
-	if numTraces == 0 && numEvents == 0 {
-		// nothing to flush
+const headerLanguages = "X-Datadog-Reported-Languages"
+
+func (w *TraceWriter) flush() {
+	if len(w.traces) == 0 && len(w.events) == 0 {
+		// nothing to do
 		return
 	}
 
 	defer timing.Since("datadog.trace_agent.trace_writer.encode_ms", time.Now())
+	defer w.resetBuffer()
 
-	atomic.AddInt64(&w.stats.Traces, int64(numTraces))
-	atomic.AddInt64(&w.stats.Events, int64(numEvents))
-	atomic.AddInt64(&w.stats.Spans, int64(w.spansInBuffer))
-
-	tracePayload := pb.TracePayload{
-		HostName:     w.hostName,
+	log.Debugf("Serializing %d traces and %d APM events.", len(w.traces), len(w.events))
+	payload := pb.TracePayload{
+		HostName:     w.hostname,
 		Env:          w.env,
 		Traces:       w.traces,
 		Transactions: w.events,
 	}
-
-	serialized, err := proto.Marshal(&tracePayload)
+	b, err := proto.Marshal(&payload)
 	if err != nil {
-		log.Errorf("Failed to serialize trace payload, data got dropped, err: %s", err)
-		w.resetBuffer()
+		log.Errorf("Failed to serialize payload, data dropped: %v", err)
 		return
 	}
-
-	atomic.AddInt64(&w.stats.BytesUncompressed, int64(len(serialized)))
-
-	encoding := "identity"
-
-	// Try to compress payload before sending
-	compressionBuffer := bytes.Buffer{}
-	gz, err := gzip.NewWriterLevel(&compressionBuffer, gzip.BestSpeed)
-	if err != nil {
-		log.Errorf("Failed to get compressor, sending uncompressed: %s", err)
-	} else {
-		_, err := gz.Write(serialized)
-		gz.Close()
-
-		if err != nil {
-			log.Errorf("Failed to compress payload, sending uncompressed: %s", err)
-		} else {
-			serialized = compressionBuffer.Bytes()
-			encoding = "gzip"
-		}
-	}
-
-	atomic.AddInt64(&w.stats.Bytes, int64(len(serialized)))
-	atomic.AddInt64(&w.stats.BytesEstimated, int64(w.bytesInBuffer))
-
-	headers := map[string]string{
-		languageHeaderKey:  strings.Join(info.Languages(), "|"),
+	p := newPayload(map[string]string{
 		"Content-Type":     "application/x-protobuf",
-		"Content-Encoding": encoding,
+		"Content-Encoding": "gzip",
+		headerLanguages:    strings.Join(info.Languages(), "|"),
+	})
+	gzipw, err := gzip.NewWriterLevel(p.body, gzip.BestSpeed)
+	if err != nil {
+		// it will never happen, unless an invalid compression is chosen;
+		// we know gzip.BestSpeed is valid.
+		log.Errorf("gzip.NewWriterLevel: %d", err)
+		return
 	}
+	gzipw.Write(b)
+	gzipw.Close()
 
-	payload := newPayload(serialized, headers)
+	atomic.AddInt64(&w.stats.BytesUncompressed, int64(len(b)))
+	atomic.AddInt64(&w.stats.BytesEstimated, int64(w.bufferedSize))
 
-	log.Debugf("Flushing traces=%v events=%v size=%d estimated=%d", len(w.traces), len(w.events), len(serialized), w.bytesInBuffer)
-	w.sender.Send(payload)
-	w.resetBuffer()
+	for _, sender := range w.senders {
+		sender.Push(p)
+	}
 }
 
-func (w *TraceWriter) resetBuffer() {
-	w.traces = w.traces[:0]
-	w.events = w.events[:0]
-	w.bytesInBuffer = 0
-	w.spansInBuffer = 0
+func (w *TraceWriter) report() {
+	metrics.Count("datadog.trace_agent.trace_writer.payloads", atomic.SwapInt64(&w.stats.Payloads, 0), nil, 1)
+	metrics.Count("datadog.trace_agent.trace_writer.bytes_uncompressed", atomic.SwapInt64(&w.stats.BytesUncompressed, 0), nil, 1)
+	metrics.Count("datadog.trace_agent.trace_writer.retries", atomic.SwapInt64(&w.stats.Retries, 0), nil, 1)
+	metrics.Count("datadog.trace_agent.trace_writer.bytes_estimated", atomic.SwapInt64(&w.stats.BytesEstimated, 0), nil, 1)
+	metrics.Count("datadog.trace_agent.trace_writer.bytes", atomic.SwapInt64(&w.stats.Bytes, 0), nil, 1)
+	metrics.Count("datadog.trace_agent.trace_writer.errors", atomic.SwapInt64(&w.stats.Errors, 0), nil, 1)
+	metrics.Count("datadog.trace_agent.trace_writer.traces", atomic.SwapInt64(&w.stats.Traces, 0), nil, 1)
+	metrics.Count("datadog.trace_agent.trace_writer.events", atomic.SwapInt64(&w.stats.Events, 0), nil, 1)
+	metrics.Count("datadog.trace_agent.trace_writer.spans", atomic.SwapInt64(&w.stats.Spans, 0), nil, 1)
 }
 
-func (w *TraceWriter) updateInfo() {
-	// TODO(gbbr): Scope these stats per endpoint (see (config.AgentConfig).AdditionalEndpoints))
-	var twInfo info.TraceWriterInfo
+var _ eventRecorder = (*TraceWriter)(nil)
 
-	// Load counters and reset them for the next flush
-	twInfo.Payloads = atomic.SwapInt64(&w.stats.Payloads, 0)
-	twInfo.Traces = atomic.SwapInt64(&w.stats.Traces, 0)
-	twInfo.Events = atomic.SwapInt64(&w.stats.Events, 0)
-	twInfo.Spans = atomic.SwapInt64(&w.stats.Spans, 0)
-	twInfo.Bytes = atomic.SwapInt64(&w.stats.Bytes, 0)
-	twInfo.BytesEstimated = atomic.SwapInt64(&w.stats.BytesEstimated, 0)
-	twInfo.BytesUncompressed = atomic.SwapInt64(&w.stats.BytesUncompressed, 0)
-	twInfo.Retries = atomic.SwapInt64(&w.stats.Retries, 0)
-	twInfo.Errors = atomic.SwapInt64(&w.stats.Errors, 0)
-	twInfo.SingleMaxSize = atomic.SwapInt64(&w.stats.SingleMaxSize, 0)
+// recordEvent implements eventRecorder.
+func (w *TraceWriter) recordEvent(t eventType, data *eventData) {
+	if data != nil {
+		metrics.Histogram("datadog.trace_agent.trace_writer.connection_fill", data.connectionFill, nil, 1)
+		metrics.Histogram("datadog.trace_agent.trace_writer.queue_fill", data.queueFill, nil, 1)
+	}
+	switch t {
+	case eventTypeRetry:
+		log.Errorf("Retrying to flush trace payload; error: %s", data.err)
+		atomic.AddInt64(&w.stats.Retries, 1)
 
-	metrics.Count("datadog.trace_agent.trace_writer.payloads", int64(twInfo.Payloads), nil, 1)
-	metrics.Count("datadog.trace_agent.trace_writer.traces", int64(twInfo.Traces), nil, 1)
-	metrics.Count("datadog.trace_agent.trace_writer.events", int64(twInfo.Events), nil, 1)
-	metrics.Count("datadog.trace_agent.trace_writer.spans", int64(twInfo.Spans), nil, 1)
-	metrics.Count("datadog.trace_agent.trace_writer.bytes", int64(twInfo.Bytes), nil, 1)
-	metrics.Count("datadog.trace_agent.trace_writer.bytes_uncompressed", int64(twInfo.BytesUncompressed), nil, 1)
-	metrics.Count("datadog.trace_agent.trace_writer.bytes_estimated", int64(twInfo.BytesEstimated), nil, 1)
-	metrics.Count("datadog.trace_agent.trace_writer.retries", int64(twInfo.Retries), nil, 1)
-	metrics.Count("datadog.trace_agent.trace_writer.errors", int64(twInfo.Errors), nil, 1)
-	metrics.Count("datadog.trace_agent.trace_writer.single_max_size", int64(twInfo.SingleMaxSize), nil, 1)
+	case eventTypeSent:
+		log.Debugf("Flushed traces to the API; time: %s, bytes: %d", data.duration, data.bytes)
+		timing.Since("datadog.trace_agent.trace_writer.flush_duration", time.Now().Add(-data.duration))
+		atomic.AddInt64(&w.stats.Bytes, int64(data.bytes))
+		atomic.AddInt64(&w.stats.Payloads, 1)
 
-	info.UpdateTraceWriterInfo(twInfo)
+	case eventTypeRejected:
+		log.Warnf("Trace writer payload rejected by edge: %v", data.err)
+		atomic.AddInt64(&w.stats.Errors, 1)
+
+	case eventTypeDropped:
+		log.Warnf("Stats writer queue full. Payload dropped (%.2fKB).", float64(data.bytes)/1024)
+		metrics.Count("datadog.trace_agent.trace_writer.dropped", 1, nil, 1)
+		metrics.Count("datadog.trace_agent.trace_writer.dropped_bytes", int64(data.bytes), nil, 1)
+	}
 }
