+++ releasenotes/notes/apm-rewrite-sender-45a220d7fffab5d9.yaml
@@ -1,0 +1,16 @@
+# Each section from every releasenote are combined when the
+# CHANGELOG.rst is rendered. So the text needs to be worded so that
+# it does not depend on any information only available in another
+# section. This may mean repeating some details, but each section
+# must be readable independently of the other.
+#
+# Each section note must be formatted as reStructuredText.
+---
+enhancements:
+  - |
+    APM: Queued sender was rewritten to improve performance around scenarios where network problems are present.
+  - |
+    APM: Code clean up around configuration and writer.
+fixes:
+  - |
+    APM: Memory spikes when retry queue grows have been fixed.
